Move dollar-single-quotes parsing to a separate module
